fixing parser

+++ src/Parser/NeuralRingerTrainerConfiguration.py
@@ -10,7 +10,7 @@ class NeuralRingerTrainerConfiguration(BaseModel):
     et_range_idx: list[int]
     eta_range_idx: list[int]
     n_initializations: int
-    pred_target_limiar: int
+    pred_target_limiar: float
     num_workers: int
     debug: bool = False
     results_folder_path: str | None = None
